+++ libs/algebra/include/nil/crypto3/algebra/fields/detail/element/fp12_fast_utils/fp12_limb_ops.hpp
@@ -1,0 +1,60 @@
+#pragma once
+
+#include <nil/crypto3/algebra/fields/detail/element/fp12_fast_utils/fp12_limb_types.hpp>
+
+namespace nil::crypto3::algebra::fields::detail::fp12_fast_utils {
+    template<typename Backend>
+    static std::array<limb, Backend::internal_limb_count> load_limbs(const Backend &backend) {
+        static_assert(Backend::limb_bits == limb_bits, "fp12 fast path expects 64-bit field limbs");
+        std::array<limb, Backend::internal_limb_count> result = {};
+        for (size_t i = 0; i < Backend::internal_limb_count; i++) {
+            result[i] = (limb)backend.limbs()[i];
+        }
+        return result;
+    }
+
+    template<size_t N>
+    inline void add_limbs_portable(limb *z, const limb *x, const limb *y) {
+        limb carry = 0u;
+        for (size_t i = 0; i < N; i++) {
+            const auto sum = (wide_limb)x[i] + y[i] + carry;
+            z[i] = (limb)sum;
+            carry = (limb)(sum >> limb_bits);
+        }
+    }
+
+    template<size_t N>
+    inline bool subtract_limbs_portable(limb *z, const limb *x, const limb *y) {
+        bool borrow = false;
+        for (size_t i = 0; i < N; i++) {
+            const limb subtrahend = y[i] + (limb)borrow;
+            const bool subtrahend_carry = subtrahend < y[i];
+            const limb current = x[i];
+            z[i] = current - subtrahend;
+            borrow = subtrahend_carry || current < subtrahend;
+        }
+        return borrow;
+    }
+
+    template<size_t N>
+    inline bool ge_modulus(const limb *x, const limb *mod) {
+        for (int i = N - 1; i >= 0; i--) {
+            if (x[i] < mod[i]) {
+                return false;
+            }
+            if (x[i] > mod[i]) {
+                return true;
+            }
+        }
+        return true;
+    }
+
+    template<size_t N>
+    inline bool ge_modulus_wide(const limb *x, const limb *mod) {
+        if (x[N] != 0u) {
+            return true;
+        }
+        return fp12_fast_utils::ge_modulus<N>(x, mod);
+    }
+
+}    // namespace nil::crypto3::algebra::fields::detail::fp12
