+++ libs/algebra/include/nil/crypto3/algebra/fields/detail/extension_params/bls12/detail/fp12_limb_types.hpp
@@ -1,23 +1,11 @@
 #pragma once
 
 #include <array>
-#include <climits>
-#include <cstddef>
-#include <cstdint>
+
+#include <nil/crypto3/algebra/fields/detail/element/fp12/fp12_limb_types.hpp>
 
 namespace nil::crypto3::algebra::fields::detail::bls12_fp12_limb_ops {
-
-    using limb = std::uint64_t;
-#if defined(__SIZEOF_INT128__)
-    using wide_limb = unsigned __int128;
-#else
-#error "bls12 fp12 limb ops require unsigned __int128 support"
-#endif
-
-    const size_t limb_bits = sizeof(limb) * CHAR_BIT;
     const size_t base_value_limb_count = 6u;
     const size_t storage_limb_count = 12u;
-
-    using limb_array = std::array<limb, storage_limb_count>;
-
+    using limb_array = std::array<fp12::limb, storage_limb_count>;
 }    // namespace nil::crypto3::algebra::fields::detail::bls12_fp12_limb_ops
